Changed all the classes to add better functionality to keep track of …

+++ src/main/java/org/bcit/comp2522/project/Projectile.java
@@ -57,7 +57,7 @@ public class Projectile extends Sprite {
   }
 
   @Override
-  public int getSize() {
+  public float getSize() {
     return size;
   }
